+++ src/CadViewerManifold.tsx
@@ -5,7 +5,7 @@ import type { ManifoldToplevel } from "manifold-3d/manifold.d.ts"
 import type React from "react"
 import { useEffect, useMemo, useState } from "react"
 import type * as THREE from "three"
-import { useThree } from "./react-three-replacement/ThreeContext"
+import { useThree } from "./react-three/ThreeContext"
 import { AnyCadComponent } from "./AnyCadComponent"
 import { CadViewerContainer } from "./CadViewerContainer"
 import { useConvertChildrenToSoup } from "./hooks/use-convert-children-to-soup"
